feat(205-01): extract useEntityDetail generic hook, convert detail hooks to thin wrappers
Create use-entity-detail.ts with parameterized fetch/draft/history logic
that accepts an EntityDetailConfig. Convert use-agent-detail,
use-skill-detail, and use-rule-detail to thin wrappers delegating to
the generic hook. Zero consumer changes required.
