+++ .planning/STATE.md
@@ -61,4 +61,4 @@ _State updated: 2026-03-25 — v6.0.0 milestone archived, state reset for next m
 
 ---
 
-_State generated from machine snapshot at 2026-03-26T15:20:54.251Z_
+_State generated from machine snapshot at 2026-03-26T15:21:53.963Z_
